pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask-cors>=5.0.1",
    "flask-login>=0.6.3",
    "flask-wtf>=1.2.2",
    "langchain>=0.3.23",
    "langchain-openai>=0.3.12",
    "neo4j>=5.28.1",
    "pillow>=11.1.0",
    "pydantic>=2.11.3",
    "pypdf>=5.4.0",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.1.0",
    "trafilatura>=2.0.0",
]
